% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solmfd_ftns.R
\name{sol_mfd_points}
\alias{sol_mfd_points}
\title{Title Solution manifold sampling}
\usage{
sol_mfd_points(
  N,
  phi,
  d,
  s,
  prior = "gaussian",
  gamma = 0.005,
  Lambda = NULL,
  tol1 = 1e-07,
  tol2 = 1e-15,
  num_iter = 1e+05,
  ...
)
}
\arguments{
\item{N}{int number of points to obtain}

\item{phi}{function target function.}

\item{d}{int dim of input of phi}

\item{s}{int dim of output of phi}

\item{prior}{str initial point prior. "gaussian" or "uniform"}

\item{gamma}{double gradient descent update step}

\item{Lambda}{matrix (s * s) positivie definite matrix.}

\item{tol1}{double tolerance for manifold convergence}

\item{tol2}{double tolerance for gradient descent algorithm}

\item{num_iter}{maximum number of iteration}

\item{...}{parameter for prior distributions}
}
\value{
matrix (n * d). Each row is one point in R^d
}
\description{
Title Solution manifold sampling
}
\examples{
set.seed(10)
N = 10
phi = function(x) {return(pnorm(2, x[[1]], x[[2]]) - pnorm(-5, x[[1]], x[[2]]) - 0.5)}
d = 2
s = 1
res_points = sol_mfd_points(N, phi, d, s, prior = "uniform")
phi(res_points[1, ]) # check == 0 for one data
plot(res_points, xlab = "mean", ylab = "sigma", type = 'o')
}
