% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solmfd_ftns.R
\name{post_density_solmfd}
\alias{post_density_solmfd}
\title{Posterior density of solution manifold points}
\usage{
post_density_solmfd(X, prob_density, points, k, h = 1, prior = "gaussian", ...)
}
\arguments{
\item{X}{matrix: (n * m). input data. n is sample number and m is dimension of data.}

\item{prob_density}{function: P(X, Z). It is in fact P(X|Z)}

\item{points}{matrix (N * d). point clouds on manifold.}

\item{k}{function: kernel function}

\item{h}{normalizing factor. Default = 1}

\item{prior}{str: type of prior distribution. Extra argument ... will be passed to a distribution parameter}

\item{...}{parameters of prior}
}
\value{
omega_i vec (N): density of each row in matrix: points's row.
}
\description{
Posterior density of solution manifold points
}
\examples{
set.seed(10)
k = get("dnorm", mode = 'function')
prob_density = function(x, theta) {return(dnorm(x, mean = theta[[1]], sd = theta[[2]]))}
n = 100
X = rnorm(n, 1.5, 3)
points = sol_mfd_points(N, phi, d, s, prior = "uniform")
res_with_density = post_density_solmfd(X, prob_density, points, k)
head(res_with_density)
}
